requirements.txt
numpy
opencv-python
pyqt5
sounddevice
matplotlib
scipy
Pillow
tensorflow
opencv-python
PyQt5>=5.6
sip